# src/main/java/com/pr0maxx/bankapplication/model/User.java
package com.pr0maxx.bankapplication.model;

import com.pr0maxx.bankapplication.model.enums.Role;
import jakarta.persistence.*;
import lombok.Getter;
import lombok.Setter;

import java.util.List;

@Entity
@Getter
@Setter
@Table(name = "users")
public class User {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @Column(unique = true, nullable = false)
    private String username; // уникальный логин
    private String password;
    private String email;

    @Enumerated(EnumType.STRING)
    @Column(nullable = false)
    private Role role; // ADMIN, USER

    @OneToMany(mappedBy = "owner", cascade = CascadeType.ALL)
    private List<Card> cards;
}